removed namespace TileLaser (TileEvent-00-09-47)
	* removed namespace TileLaser
	* Tagging TileEvent-00-09-47

2015-05-12  Marius Cornelis van Woerden <mvanwoer@cern.ch>
        * Tagging TileEvent-00-09-46
        * solved pedestal bug

2015-04-30  Sanya Solodkov  <solodkov@pcatl08.cern.ch>

	* Tagging TileEvent-00-09-45
	* TileEvent/TileCell.h: Protect against undefined behavior in bit shift

2015-03-18  scott snyder  <snyder@bnl.gov>

	* Tagging TileEvent-00-09-43.
	* src/TileCell.cxx: Protect against undefined behavior in bit
	shift.

+++ TileCalorimeter/TileEvent/src/TileCell.cxx
@@ -67,7 +67,7 @@ TileCell::TileCell(const CaloDetDescrElement* const & caloDDE,
                    int qual1, int qual2, int qbit1, int qbit2, 
                    int gain1, int gain2)
   : CaloCell(caloDDE,ene1+ene2,(time1+time2)/2.0, 0, 0,
-             (CaloGain::CaloGain) ( 0xFFFFFFF0 | (gain2 << 2) | (gain1 & 3) ))
+             (CaloGain::CaloGain) ( 0xFFFFFFF0 | (static_cast<unsigned int>(gain2) << 2) | (gain1 & 3) ))
   , m_eneDiff(ene1-ene2)
   , m_timeDiff((time1-time2)/2.)
 {
